Fix Subfolder trees #5
