+++ src/utils/github/operations/index.ts
@@ -1,0 +1,4 @@
+export { processRepo } from './repo';
+export { processPath } from './path';
+export { processFile } from './file';
+export { processDirectoryContents } from './directory';
